refactor(intune): align Win32 module names and surface with the family (#357)
Pure rename plus encapsulation, no behavior change (2246 tests
unchanged before and after):

- signals.rs -> rules.rs (record classification) and rules.rs ->
  findings.rs (derive_findings), matching microsoft_store and the rest
  of the family where 'rules' means classification and 'findings' means
  findings.
- private modules with one curated pub use surface: every public name
  is now reachable at exactly one path (win32::Name); the previous
  pub-mod-everything layout published each item at two paths.

Reference sweep per the no-semantic-search checklist: direct imports
(reducer.rs super::signals), type-level references (none external),
string/doc references (mod.rs table and ownership note), re-exports
(mod.rs curated list), and tests (integration tests import only the
curated surface; no test named the old paths).
